refactor: SurveyAnswer 도메인의 Dto 변환 클래스 Factory -> Mapper 로 이름 변경
